app: Refactor webhooks into module
